line_provider/pyproject.toml
[tool.poetry]
name = "line-provider"
version = "0.1.0"
description = ""
authors = ["Danila Redikultsev <danila.redikulcev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
redis = {extras = ["async"], version = "^5.2.1"}
pydantic-settings = "^2.7.0"
fastapi-cache2 = "^0.2.2"
httpx = "^0.28.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
